[DNM] jobs: prevent the race that can cause multiple resumers to run concurrently for the same job
A resumer executes a job. Resumers keep internal state to prevent
concurrent execution of resumers for the same job. The state is
cleared when a resumer finishes or the job is canceled/paused in
a concurrent goroutine "cancel-loop". The race between a resumer
and cancel-loop to set and clear the internal state can result
in concurrent execution of multiple resumers for the same job,
which can lead to unexpected results. This commit prevents the
race to happen. Although the race is unlikely to happen in
production, this commit will prevent flaky tests in jobs due
to that race.

Release note: None